The Backend Code Of LostAndFoundSystem

+++ src/main/java/org/example/service/impl/UserServiceImpl.java
@@ -24,26 +24,30 @@ public class UserServiceImpl implements UserService {
     @Autowired
     private AliOSSUtils aliOSSUtils;
 
+    @Override
     public List<User> listUser() {
         return userMapper.selectUser("Alice");
     }
 
+    @Override
     public String home(String username) {
-        return userMapper.getAvatar(username);
+        return userMapper.getAvatarByUsername(username);
     }
 
+    @Override
     public User personalhome(String username) {
         User user;
         boolean exists = Boolean.TRUE.equals(redisTemplate.hasKey(username));
         if (exists == true) {
             user = redisTemplate.opsForValue().get(username);
         } else {
-            user = userMapper.getUser(username);
+            user = userMapper.getUserByUsername(username);
             redisTemplate.opsForValue().set(username, user, 43200000, TimeUnit.MILLISECONDS);
         }
         return user;
     }
 
+    @Override
     public void update(String username, String newUsername, String newPassword, String newPhone, MultipartFile newAvatar) throws IOException {
         String newAvatarUrl = null;
         String oldAvatarUrl = this.personalhome(username).getAvatar();
@@ -51,12 +55,29 @@ public class UserServiceImpl implements UserService {
             newAvatarUrl = aliOSSUtils.upload(newAvatar);
             userMapper.updateUser(username, newUsername, newPassword, newPhone, newAvatarUrl);
             redisTemplate.delete(username);
-            aliOSSUtils.delete(oldAvatarUrl);
+            if (!oldAvatarUrl.equals("https://lost-and-found-system-bucket.oss-cn-hangzhou.aliyuncs.com/avatar.jpg")) {
+                aliOSSUtils.delete(oldAvatarUrl);
+            }
         } catch (Exception e) {
             if (newAvatarUrl != null) {
                 aliOSSUtils.delete(newAvatarUrl);
             }
             throw e;
+        }
+    }
+
+    @Override
+    public void logout(String username) {
+        redisTemplate.delete(username);
+    }
+
+    @Override
+    public void cancel(String username) {
+        redisTemplate.delete(username);
+        String avatar = userMapper.getAvatarByUsername(username);
+        userMapper.deleteUserByUsername(username);
+        if (!avatar.equals("https://lost-and-found-system-bucket.oss-cn-hangzhou.aliyuncs.com/avatar.jpg")) {
+            aliOSSUtils.delete(avatar);
         }
     }
 }
